+++ src/unstruc/block.cpp
@@ -10,15 +10,15 @@
 
 namespace unstruc {
 
-Block::Block(int s1, int s2, int s3) : size1(s1), size2(s2), size3(s3) {
+Block::Block(size_t s1, size_t s2, size_t s3) : size1(s1), size2(s2), size3(s3) {
 	points = std::vector <Point> (size1*size2*size3);
 };
 
-Point Block::at(int i, int j, int k) {
+Point Block::at(size_t i, size_t j, size_t k) {
 	return points[i*(size2*size3) + j*size3 + k];
 };
-double * Block::at_ref(int i, int j, int k, int l) {
-	int ii = i*(size2*size3) + j*size3 + k;
+double * Block::at_ref(size_t i, size_t j, size_t k, size_t l) {
+	size_t ii = i*(size2*size3) + j*size3 + k;
 	switch (l) {
 		case 0:
 			return &(points[ii].x);
@@ -30,32 +30,32 @@ double * Block::at_ref(int i, int j, int k, int l) {
 			return NULL;
 	}
 };
-int Block::index(int i, int j, int k) {
+size_t Block::index(size_t i, size_t j, size_t k) {
 	return i*(size2*size3) + j*size3 + k;
 };
 
 Grid MultiBlock::to_grid() {
 	Grid grid(3);
-	int offset = 0;
+	size_t offset = 0;
 	std::stringstream ss;
 	std::cerr << "Converting to unstructured grid" << std::endl;
-	int si,sj,sk;
-	for (int ib = 0; ib < blocks.size(); ib++) {
+	size_t si,sj,sk;
+	for (size_t ib = 0; ib < blocks.size(); ib++) {
 		Block& blk = blocks[ib];
 		si = blk.size1;
 		sj = blk.size2;
 		sk = blk.size3;
-		for (int i = 0; i < si; i++)
-			for (int j = 0; j < sj; j++)
-				for (int k = 0; k < sk; k++)
+		for (size_t i = 0; i < si; i++)
+			for (size_t j = 0; j < sj; j++)
+				for (size_t k = 0; k < sk; k++)
 					grid.points.push_back(blk.at(i,j,k));
 		ss.str("");
 		ss.clear();
 		ss << "Block" << ib+1;
 		grid.names.push_back(Name(3,ss.str()));
-		for (int i = 0; i < si-1; i++) {
-			for (int j = 0; j < sj-1; j++) {
-				for (int k = 0; k < sk-1; k++) {
+		for (size_t i = 0; i < si-1; i++) {
+			for (size_t j = 0; j < sj-1; j++) {
+				for (size_t k = 0; k < sk-1; k++) {
 					Element e = Element(Shape::Hexa);
 					e.points[0] = offset+blk.index(i,j,k);
 					e.points[1] = offset+blk.index(i+1,j,k);
@@ -74,9 +74,9 @@ Grid MultiBlock::to_grid() {
 		ss.clear();
 		ss << "Block" << ib+1 << " FaceI1";
 		grid.names.push_back(Name(2,ss.str()));
-		int i = 0;
-		for (int j = 0; j < sj-1; j++) {
-			for (int k = 0; k < sk-1; k++) {
+		size_t i = 0;
+		for (size_t j = 0; j < sj-1; j++) {
+			for (size_t k = 0; k < sk-1; k++) {
 				Element e = Element(Shape::Quad);
 				e.points[0] = offset+blk.index(i,j,k);
 				e.points[1] = offset+blk.index(i,j+1,k);
@@ -91,8 +91,8 @@ Grid MultiBlock::to_grid() {
 		ss << "Block" << ib+1 << " FaceI2";
 		grid.names.push_back(Name(2,ss.str()));
 		i = si-1;
-		for (int j = 0; j < sj-1; j++) {
-			for (int k = 0; k < sk-1; k++) {
+		for (size_t j = 0; j < sj-1; j++) {
+			for (size_t k = 0; k < sk-1; k++) {
 				Element e = Element(Shape::Quad);
 				e.points[0] = offset+blk.index(i,j,k);
 				e.points[1] = offset+blk.index(i,j+1,k);
@@ -106,9 +106,9 @@ Grid MultiBlock::to_grid() {
 		ss.clear();
 		ss << "Block" << ib+1 << " FaceJ1";
 		grid.names.push_back(Name(2,ss.str()));
-		int j = 0;
-		for (int i = 0; i < si-1; i++) {
-			for (int k = 0; k < sk-1; k++) {
+		size_t j = 0;
+		for (size_t i = 0; i < si-1; i++) {
+			for (size_t k = 0; k < sk-1; k++) {
 				Element e = Element(Shape::Quad);
 				e.points[0] = offset+blk.index(i,j,k);
 				e.points[1] = offset+blk.index(i+1,j,k);
@@ -123,8 +123,8 @@ Grid MultiBlock::to_grid() {
 		ss << "Block" << ib+1 << " FaceJ2";
 		grid.names.push_back(Name(2,ss.str()));
 		j = sj-1;
-		for (int i = 0; i < si-1; i++) {
-			for (int k = 0; k < sk-1; k++) {
+		for (size_t i = 0; i < si-1; i++) {
+			for (size_t k = 0; k < sk-1; k++) {
 				Element e = Element(Shape::Quad);
 				e.points[0] = offset+blk.index(i,j,k);
 				e.points[1] = offset+blk.index(i+1,j,k);
@@ -138,9 +138,9 @@ Grid MultiBlock::to_grid() {
 		ss.clear();
 		ss << "Block" << ib+1 << " FaceK1";
 		grid.names.push_back(Name(2,ss.str()));
-		int k = 0;
-		for (int i = 0; i < si-1; i++) {
-			for (int j = 0; j < sj-1; j++) {
+		size_t k = 0;
+		for (size_t i = 0; i < si-1; i++) {
+			for (size_t j = 0; j < sj-1; j++) {
 				Element e = Element(Shape::Quad);
 				e.points[0] = offset+blk.index(i,j,k);
 				e.points[1] = offset+blk.index(i+1,j,k);
@@ -155,8 +155,8 @@ Grid MultiBlock::to_grid() {
 		ss << "Block" << ib+1 << " FaceK2";
 		grid.names.push_back(Name(2,ss.str()));
 		k = sk-1;
-		for (int i = 0; i < si-1; i++) {
-			for (int j = 0; j < sj-1; j++) {
+		for (size_t i = 0; i < si-1; i++) {
+			for (size_t j = 0; j < sj-1; j++) {
 				Element e = Element(Shape::Quad);
 				e.points[0] = offset+blk.index(i,j,k);
 				e.points[1] = offset+blk.index(i+1,j,k);
